+++ src/atom/feeds.rs
@@ -1,19 +1,15 @@
 use ammonia::clean_text;
-use chrono::{DateTime, NaiveDateTime, Utc};
 
 use crate::activitypub::identifiers::{local_actor_id, local_object_id};
 use crate::config::Instance;
 use crate::models::posts::types::Post;
 use crate::models::profiles::types::DbActorProfile;
-use crate::utils::html::clean_html_all;
+use crate::utils::{
+    datetime::get_min_datetime,
+    html::clean_html_all,
+};
 
 const ENTRY_TITLE_MAX_LENGTH: usize = 75;
-
-fn get_min_datetime() -> DateTime<Utc> {
-    let native = NaiveDateTime::from_timestamp_opt(0, 0)
-        .expect("0 should be a valid argument");
-    DateTime::from_utc(native, Utc)
-}
 
 fn make_entry(
     instance_url: &str,