+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Save sizes of media attachments and other files to database.
+- Added `import-emoji` command.
 
 ### Security